+++ NEWS.md
@@ -1,6 +1,9 @@
 # odytools 0.7.5.9000
 
--   Improved **ody_qlq_c30_v3** with new criteria for the calculation of scales when some items are missing, and added the capability for delta calculation.
+-   Improved **ody_qlq_c30_v3**:
+    - The functions checks if the items have the expected range of values.
+    - New criteria for the calculation of scales when some items are missing.
+    - Added the capability for delta calculation.
 -   Added **ody_gt2image** to convert a GT table into an image.
 -   When accessing a project (both plain and REDCap-based), the existence and update status of the renv.lock file and a git repo are now checked.
 -   When rendering a report via the render chunk, the lockfile is now updated if the report is successfully rendered.